shop/admin.py
from django.contrib import admin
from .models import Category, Product, Comment
from parler.admin import TranslatableAdmin

 
@admin.register(Category)
class CategoryAdmin(TranslatableAdmin):
    list_display = ['name', 'slug']
 
    def get_prepopulated_fields(self, request, obj=None):
        return {'slug': ('name',)}


@admin.register(Product)
class ProductAdmin(TranslatableAdmin):
    list_display = ['name', 'slug', 'price', 'available', 'created', 'updated']
    list_filter = ['available', 'created', 'updated']
    list_editable = ['price', 'available']
     
    def get_prepopulated_fields(self, request, obj=None):
        return {'slug': ('name',)}


@admin.register(Comment)
class CommentAdmin(admin.ModelAdmin):
    list_display = ('id_user', 'user', 'name', 'product', 'created', 'active')
    list_filter = ('active', 'created', 'updated')
    search_fields = ('user', 'name', 'body')


