requirements.txt
-f /usr/share/pip-wheels
braintree==3.49.0
cairocffi==0.9.0
CairoSVG==2.2.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.5.0
Django==2.1.3
django-localflavor==2.1
django-parler==1.9.2
django-rosetta==0.9.0
html5lib==1.0.1
idna==2.7
Pillow==5.3.0
polib==1.1.0
psutil==5.4.8
pycparser==2.19
Pyphen==0.9.5
pytz==2018.7
redis==3.0.1
redislite==3.2.311
requests==2.20.1
six==1.11.0
sorl-thumbnail==12.5.0
tinycss2==0.6.1
urllib3==1.24.1
WeasyPrint==43
webencodings==0.5.1
